Add files via upload
An organization loses its customers to its competition for various reasons. What mainly attracts customers to go ahead with a shift is attributed to the price of the product and its quality. This is true for the Telecom industry as well. A churn of 10 to 60 percent in any company is a big number that can affect the company’s overall growth.

Churn is the percentage of customers who stop using a company's product or service during a particular period of time. Churn category refers to the reason why a customer stops using a company's service, while churn reason refers to the specific reason for a customer to churn. In the telecom industry, multiclass classification is used to predict a customer's churn category and churn reason.